backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
python-dotenv==1.1.0
pydantic==2.11.3
pydantic-settings==2.9.1
google-genai==1.14.0
tavily-python==0.5.0
firecrawl-py==1.16.0
docxtpl==0.19.0
supabase==2.15.2
PyJWT[crypto]>=2.9.0
sse-starlette==2.3.3
httpx==0.28.1
python-docx==1.1.2
